requirements.txt
streamlit
python-docx
fountain
openai
elevenlabs
python-dotenv